+++ extensions/toothpick/src/core/devices/devices.service.ts
@@ -1,5 +1,6 @@
 import ApplescriptDevicesService from "./devices.service.applescript";
 import BlueutilDevicesService from "./devices.service.blueutil";
+import WindowsDevicesService from "./devices.service.windows";
 import { Device } from "./devices.model";
 
 export interface DevicesService {
@@ -15,14 +16,18 @@ let devicesService: DevicesService;
 export function getDevicesService(
   serviceType?: ServiceType | ExtensionPreferences["bluetoothBackend"],
 ): DevicesService {
-  if (serviceType && currentServiceType !== serviceType) {
-    currentServiceType = serviceType;
-    switch (serviceType) {
+  const requestedService = process.platform === "win32" ? ServiceType.Windows : serviceType;
+  if (requestedService && currentServiceType !== requestedService) {
+    currentServiceType = requestedService;
+    switch (requestedService) {
       case ServiceType.AppleScript:
         devicesService = new ApplescriptDevicesService();
         break;
       case ServiceType.Blueutil:
         devicesService = new BlueutilDevicesService();
+        break;
+      case ServiceType.Windows:
+        devicesService = new WindowsDevicesService();
         break;
     }
   }
@@ -32,4 +37,5 @@ export function getDevicesService(
 enum ServiceType {
   AppleScript = "applescript",
   Blueutil = "blueutil",
+  Windows = "windows",
 }